skills/game-system-designer/requirements.txt
requests>=2.28.0
Pillow>=9.0.0
openpyxl>=3.0.0
